Subida ejercicio evaluable

+++ Ejercicios/monumentos/src/main/java/com/salesianostriana/dam/monumentos/service/MonumentoService.java
@@ -5,6 +5,7 @@ import com.salesianostriana.dam.monumentos.repository.MonumentoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -13,16 +14,21 @@ public class MonumentoService {
     @Autowired
     private MonumentoRepository monumentoRepository;
 
-    public Optional<Monumento> obtenerPorId(long id){
+    public List<Monumento> ObtenerTodos(){
+        return monumentoRepository.findAll();
+    }
+
+    public Optional<Monumento> obtenerPorId(Long id){
         return monumentoRepository.findById(id);
     }
     public Monumento guardar(Monumento monumento){ //Guarda y actualiza
         return monumentoRepository.save(monumento);
     }
-    public Monumento modificarMonumento(Long id, Monumento monumento){
+    public Monumento modificarMonumento(Long id){
         return null;
     }
-    public void eliminarMonumento(long id){
-        monumentoRepository.deleteById(id);
+    public void eliminarMonumento(Long id){
+        if(obtenerPorId(id).isPresent())
+            monumentoRepository.deleteById(id);
     }
 }
